Fix Question3.c segmentation fault for large prime numbers
The factor recursive function was replaced by a normal one. This is
because a large prime number n previously required n recursive calls
to the function, smashing the stack.

One alternative was to initially check if n was prime. However, for
non-prime numbers this would require double the amount of prime
checking.